chore(ui): rename utils/notifications.ts to utils/toasts.ts to distinguish
