+++ ui/src/components/ButtonFollow.vue
@@ -21,7 +21,7 @@
 
 <script setup lang="ts">
 import { PROFILE_FIELDS } from "@/types/types";
-import { showError, showMessage } from "@/utils/notification";
+import { showError, showMessage } from "@/utils/toasts";
 import { AgentPubKey } from "@holochain/client";
 import { ComputedRef, inject, onMounted, PropType, ref } from "vue";
 import { QBtn, QIcon } from "quasar";
