+++ ui/src/pages/MewYarn.vue
@@ -60,7 +60,7 @@ import { ComputedRef, computed, inject, ref, watch } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { AppAgentClient } from "@holochain/client";
 import isEqual from "lodash/isEqual";
-import { showMessage } from "@/utils/notification";
+import { showMessage } from "@/utils/toasts";
 
 const client = (inject("client") as ComputedRef<AppAgentClient>).value;
 const route = useRoute();
